factor out separator collapsing
